requirements.txt
biopython==1.65
decorator==4.0.4
enum34==1.1.1
Flask==0.10.1
Flask-Assets==0.11
funcsigs==0.4
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
matplotlib==1.4.3
mock==1.3.0
networkx==1.10
nose==1.3.7
numpy==1.10.1
pathlib==1.0.1
Pattern==2.6
pbr==1.8.1
psycopg2>=2.6.1
pydot2==1.0.33
pyparsing==2.0.3
pyScss==1.3.4
python-dateutil==2.4.2
python-memcached==1.57
pytz==2015.6
redis==2.10.5
scipy==0.18.0
six==1.10.0
webassets==0.11.1
Werkzeug==0.10.4
